增加tiled旋转的逻辑

+++ Assets/Script/DataMgr.cs
@@ -6,23 +6,23 @@ public class DataMgr
 {
     private readonly Vector2Int mapAreaSize = new Vector2Int(10, 10);
 
-    private List<short[,]> terrain = null;
+    private List<uint[,]> terrain = null;
 
     public void DeserializeTerrainData(Object obj)
     {
         TextAsset asset = obj as TextAsset;
         if (asset != null)
         {
-            terrain = ByteUtils.DeSerializable(asset.bytes) as List<short[,]>;
+            terrain = ByteUtils.DeSerializable(asset.bytes) as List<uint[,]>;
         }
     }
 
-    public int GetUVIndexByPosition(int index,Vector2 position)
+    public uint GetUVIndexByPosition(int index,Vector2 position)
     {
         Vector2 pos = position + (mapAreaSize/2);
         int pos_x = Mathf.FloorToInt(pos.x/2);
         int pos_y = Mathf.FloorToInt(pos.y/2);
-        int uvIndex = terrain[index][pos_x, pos_y];
+        uint uvIndex = terrain[index][pos_x, pos_y];
         return uvIndex;
     }
 }
